- Salvando cidade no banco de dados

+++ app/src/main/java/com/nevesrafael/tenki/model/CityApiResponse.kt
@@ -3,8 +3,8 @@ package com.nevesrafael.tenki.model
 data class CityApiResponse(
 
     val name: String,
-    val lat: Long,
-    val lon: Long,
+    val lat: Double,
+    val lon: Double,
     val country: String,
     val state: String
 )
